[Bronze V] Title: 이상한 기호, Time: 68 ms, Memory: 11512 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,8 +5,8 @@ public class Main{
 	public static void main(String[] args) throws IOException{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		String [] inputData = br.readLine().trim().split(" ");
-		int a = Integer.parseInt(inputData[0]);
-		int b = Integer.parseInt(inputData[1]);
+		long a = Long.parseLong(inputData[0]);
+		long b = Long.parseLong(inputData[1]);
 		
 		System.out.print((a+b)*(a-b));
 	}
